Add user validation

+++ routes/web.php
@@ -16,14 +16,14 @@ $router->get('/', function () use ($router) {
 });
 
 /** Messages **/
-$router->get('/messages/{threadId}', 'MessageController@get');
+$router->get('/messages/{threadId}[/{userId}]', 'MessageController@get');
 $router->post('/message/{threadId}', 'MessageController@add');
 
 /** Threads **/
 $router->get('/threads/{userId}', 'ThreadController@get');
 $router->post('/thread', 'ThreadController@add');
-$router->put('/thread/{threadId}', 'ThreadController@update');
-$router->delete('/thread/{threadId}', 'ThreadController@delete');
+$router->put('/thread/{threadId}[/{userId}]', 'ThreadController@update');
+$router->delete('/thread/{threadId}[/{userId}]', 'ThreadController@delete');
 
-$router->post('/thread/users/{threadId}', 'ThreadController@addUsers');
-$router->delete('/thread/users/{threadId}', 'ThreadController@removeUsers');
+$router->post('/thread/users/{threadId}[/{userId}]', 'ThreadController@addUsers');
+$router->delete('/thread/users/{threadId}[/{userId}]', 'ThreadController@removeUsers');
